backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.6.0

# CORS
fastapi-cors==0.0.6

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
faker==30.3.0

# Development
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.11.2

# Documentation
pydantic==2.9.2